+++ Code Design and Data Structures/Assignments/aieBootstrap-master/Simon Says/MenuButton.h
@@ -6,7 +6,7 @@
 class MenuButton
 {
 public:
-	MenuButton(const char* buttonText, float x, float y, float width, float height, int r, int g, int b);
+	MenuButton(const char* buttonText, float x, float y, float width, float height, float r, float g, float b);
 	~MenuButton();
 
 	void Draw(aie::Renderer2D* renderer, float width, float height);
